Fix isomorphism check to handle other A values
Previously the "A" coefficient (penalty associated with HA term in
objective) was specified in the code as a variable, but the check for
an isomorphism based on the ground state energy assumed that A=1.
Resolved this by moving the A and B coefficients into the API for the
various problem definition/solution functions and using the value of A
in the check on the energy value.

Also replaced the equality check with np.isclose in attempt to be more
robust to floating point comparison issues.