Simplify code by removing A and B coefficients
The A and B values are the penalty coefficients associated with the HA
and HB terms in the objective function.  Since current code takes
these both to be 1, which seems to work well, it seems better to
simply remove them from the code, which simplifies the interfaces for
several of the functions.